requirements.txt
python-dotenv
requests
pymongo
APScheduler
pyspark==3.4.4
delta-spark==2.4.0
influxdb-client